refactor(api-file-manager): delete barrel abstractions.ts, direct imports everywhere
Internal consumers import directly from abstractions/<Name>.js.
External consumers (s3, server packages) import from
exports/api/file-manager/assetDelivery.js. Raw IFoo interface usages
in privateFiles replaced with namespace types (AssetOutputStrategy.Interface).
