+++ packages/api-file-manager-s3/src/assetDelivery/s3/S3AssetResolver.ts
@@ -1,9 +1,9 @@
 import type { S3 } from "@webiny/aws-sdk/client-s3/index.js";
 import type { AssetRequest, AssetResolver } from "@webiny/api-file-manager";
 import { Asset } from "@webiny/api-file-manager";
-import { AssetResolver as AssetResolverAbstraction } from "@webiny/api-file-manager/features/assetDelivery/abstractions.js";
+import { AssetResolver as AssetResolverAbstraction } from "@webiny/api-file-manager/exports/api/file-manager/assetDelivery.js";
 import { GlobalKeyValueStore } from "@webiny/api-core/features/keyValueStore/index.js";
-import { ObjectKey } from "@webiny/api-file-manager/features/assetDelivery/ObjectKey/index.js";
+import { ObjectKey } from "@webiny/api-file-manager/exports/api/file-manager/assetDelivery.js";
 import { S3ContentsReader } from "~/assetDelivery/index.js";
 import { S3Client, S3Bucket } from "~/assetDelivery/abstractions.js";
