+++ packages/api-file-manager/src/features/assetDelivery/NullAssetOutputStrategy.ts
@@ -1,5 +1,5 @@
 import { NullAssetReply } from "./NullAssetReply.js";
-import { AssetOutputStrategy } from "./abstractions.js";
+import { AssetOutputStrategy } from "./abstractions/AssetOutputStrategy.js";
 
 class NullAssetOutputStrategyImpl implements AssetOutputStrategy.Interface {
     async output(): Promise<AssetOutputStrategy.AssetReply> {
