--- packages/api-file-manager/src/features/assetDelivery/abstractions.ts
@@ -1,19 +1,0 @@
-export {
-    AssetRequestResolver,
-    type IAssetRequestResolver
-} from "./abstractions/AssetRequestResolver.js";
-export { AssetResolver, type IAssetResolver } from "./abstractions/AssetResolver.js";
-export { AssetProcessor, type IAssetProcessor } from "./abstractions/AssetProcessor.js";
-export {
-    AssetOutputStrategy,
-    type IAssetOutputStrategy
-} from "./abstractions/AssetOutputStrategy.js";
-export {
-    AssetTransformationStrategy,
-    type IAssetTransformationStrategy
-} from "./abstractions/AssetTransformationStrategy.js";
-export {
-    AssetContentsReader,
-    type IAssetContentsReader
-} from "./abstractions/AssetContentsReader.js";
-export { AssetAuthorizer, type IAssetAuthorizer } from "./abstractions/AssetAuthorizer.js";
