+++ packages/api-file-manager/src/features/assetDelivery/privateFiles/NotAuthorizedOutputStrategy.ts
@@ -1,8 +1,8 @@
 import type { AssetReply } from "~/delivery/AssetDelivery/abstractions/AssetReply.js";
 import { NotAuthorizedAssetReply } from "./NotAuthorizedAssetReply.js";
-import type { IAssetOutputStrategy } from "../abstractions.js";
+import { AssetOutputStrategy } from "../abstractions/AssetOutputStrategy.js";
 
-export class NotAuthorizedOutputStrategy implements IAssetOutputStrategy {
+export class NotAuthorizedOutputStrategy implements AssetOutputStrategy.Interface {
     async output(): Promise<AssetReply> {
         return new NotAuthorizedAssetReply();
     }
