+++ packages/api-file-manager/src/features/assetDelivery/privateFiles/RedirectToPrivateUrlOutputStrategy.ts
@@ -2,9 +2,9 @@ import type { Asset } from "~/delivery/AssetDelivery/Asset.js";
 import type { AssetRequest } from "~/delivery/AssetDelivery/AssetRequest.js";
 import { AssetReply } from "~/delivery/AssetDelivery/abstractions/AssetReply.js";
 import { ResponseHeaders } from "@webiny/handler";
-import type { IAssetOutputStrategy } from "../abstractions.js";
+import { AssetOutputStrategy } from "../abstractions/AssetOutputStrategy.js";
 
-export class RedirectToPrivateUrlOutputStrategy implements IAssetOutputStrategy {
+export class RedirectToPrivateUrlOutputStrategy implements AssetOutputStrategy.Interface {
     private assetRequest: AssetRequest;
 
     constructor(assetRequest: AssetRequest) {
